#173: Move data for sections out of the component files
